When the user does not fill in the fields, the property is saved correctly.

+++ app/src/main/java/com/example/realestatemanager/data/local/property/PropertyEntity.kt
@@ -10,7 +10,7 @@ import kotlinx.android.parcel.Parcelize
 data class PropertyEntity(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val type: String?,
-    val price: Double?,
+    val price: Long?,
     val surface: Int?,
     val room: Int?,
     val image: String?,
